Backend/requirements.txt
# Flask e extensões
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3

# Banco de dados
SQLAlchemy==2.0.41
greenlet==3.2.3

# Requisições HTTP
requests==2.32.4
urllib3==2.5.0

# Processamento de dados
numpy==2.3.0
pandas==2.3.0

# Geoprocessamento
pyogrio==0.11.0
pyproj==3.7.1
geopandas

# Utilitários
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Validação e formatação
click==8.2.1
Jinja2==3.1.6
MarkupSafe==3.0.2

# Segurança e certificados
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10

# Outros
blinker==1.9.0
itsdangerous==2.2.0
packaging==25.0
six==1.17.0
typing_extensions==4.14.0

# Desenvolvimento (opcional)
python-dotenv==1.0.0
